server/middleware/authMiddlware.py
 # middlewares/auth_middleware.py

from functools import wraps
from flask import request, jsonify
import jwt
import os
from dotenv import load_dotenv

load_dotenv()
SECRET = os.getenv("ACCESS_TOKEN_SECRET", "dev-secret")

def auth_required(f):
    @wraps(f)
    def decorated(*args, **kwargs):
        token = request.headers.get('Authorization')
        if not token:
            return jsonify({"error": "Access denied"}), 401
        try:
            token = token.replace("Bearer ", "")
            decoded = jwt.decode(token, SECRET, algorithms=["HS256"])
            request.user = decoded
        except jwt.ExpiredSignatureError:
            return jsonify({"error": "Token expired"}), 401
        except jwt.InvalidTokenError:
            return jsonify({"error": "Invalid token"}), 400
        return f(*args, **kwargs)
    return decorated
